p2psafety_django/requirements.txt
Django==1.6
django-annoying==0.7.9
django-tastypie==0.9.16
south==0.8
python-social-auth==0.1.20
psycopg2==2.5
mock==1.0.1
git+git://github.com/alekzvik/factory_boy.git
